requirements.txt
accelerate>=0.21.0
chromadb
cmake
crewai
ctransformers
diffusers
faiss
faiss-cpu
gradio
huggingface-hub
langchain
langchain-community
llama-cpp-python
llama_index
llama-index-embeddings-huggingface
llama-index-llms-llama-cpp
llama-index-llms-langchain
llama-index-vector-stores-faiss
modin[all]
numpy
openai
pymupdf
pypdf
PyPDF2
python-dotenv
qdrant-client
sentence-transformers
streamlit
streamlit_chat
tensorflow
tiktoken
torch
torchaudio
torchvision
transformers